feat: refactor code     - seperate middleware from main package to its own middleware package     - seperate router from main package to its own router package
